// types/siteSettings.ts
export type MenuType = {
    _id: string;
    id: number;
    title: string;
    name: string;
    url: string;
    type: string;
    buttonLink: string;
    buttonLinks: string;
    buttonUrl: string;
    buttonText: string;
  };

  export type SliderType = {
    _id: string;
    id: number;
    title: string;
    name: string;
    logo: string;
    banner: string;
  };

  export type SettingsType = {
    _id: string;   
    title: string;
    name: string;
    description: string;
    logo: string;
  };
   